hierarchical_clustering and biostatistics parts are separated into 2 files
